Workaround ROS2 not working if there's no network connection
:Release Notes:
Workaround ROS2 not working if there's no network connection

:Detailed Notes:
We (Brian, Martins and I) noticed that ROS2 doesn't work if its
underlying DDS doesn't find all network interfaces that have a MAC
with a valid IP assigned.

As a workaround, we wait until network stack is fully up so it
doesn't undo our forced IP assignment for each of the network
interfaces with MACs.

:Testing Performed:
Tested on rpi locally

:QA Notes:
N/A

:Issues Addressed:

Change-Id: Idd0a0156cfd2563a7038a48493a33d54cf455e9c